AI-Powered Meeting/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-generativeai==0.3.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
python-multipart==0.0.6